Remove spec_api_helper and create publish_version helper.
