+++ spec/support/api_spec_helpers.rb
@@ -2,4 +2,23 @@ module ApiSpecHelpers
   def json_body
     JSON.parse(response.body)
   end
+
+  def signature(resource)
+    resource.except('created_at', 'updated_at')
+  end
+
+  def error_404
+    {
+       'error_messages' => ['Resource does not exist'],
+       'error_code' => 'NOT_FOUND'
+    }
+  end
+
+  def publish_version(cookbook, version)
+    create(
+      :cookbook_version,
+      cookbook: cookbook,
+      version: version
+    )
+  end
 end
